added StrategyRetriever class . Fully handled ATM_Device class . Added sample code in Program.cs

+++ ATM/User/Interfaces/IUserFactory.cs
@@ -8,6 +8,6 @@ namespace ATM.User.Interfaces
 {
     internal interface IUserFactory
     {
-        BaseUser CreateUser(UserType userType, string userName, double moneyInAccount);
+        BaseUser CreateUser(UserType userType, string userName, decimal moneyInAccount);
     }
 }
